refactor: structure, general moments & future api contract
