+++ packages/frontend/src/features/quiz/pages/categories-page/categories-page.component.html
@@ -3,5 +3,5 @@
     <h2 class="h2">Categories</h2>
   </header>
 
-  <app-category-card-list class="categories-page__category-card-list" [categories]="categories" />
+  <app-category-card-list class="category-page__category-card-list" [categories]="categories" />
 </section>
