+++ packages/frontend/src/features/quiz/ui/category-card-list/category-card-list.component.html
@@ -4,7 +4,7 @@
   @for (listItem of list; track listItem.id) {
     <li>
       <app-category-card
-        [heading]="listItem.topic"
+        [heading]="listItem.name"
         [description]="listItem.description"
         [topicsCount]="listItem.topicsCount"
         [topicsCompleteCount]="listItem.topicsCompleteCount"
