+++ packages/frontend/src/features/quiz/ui/category-card-list/category-card-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, input } from '@angular/core';
 
-import { CategoryCardComponent } from '../category-card/category-card.component';
+import { CategoryCardComponent } from './components';
 
 import type { CategoryCardListItem } from './category-card-list.types';
