+++ packages/frontend/src/features/quiz/ui/category-card-list/category-card-list.types.ts
@@ -1,6 +1,6 @@
 interface CategoryCardListItem {
   id: string;
-  topic: string;
+  name: string;
   description: string;
   topicsCount: number;
   topicsCompleteCount: number;
