+++ packages/frontend/src/features/quiz/ui/category-card-list/components/category-card/category-card.component.scss
