+++ packages/frontend/src/features/quiz/ui/category-card-list/components/index.ts
@@ -1,0 +1,1 @@
+export { CategoryCardComponent } from './category-card/category-card.component';
