+++ packages/frontend/src/features/quiz/ui/topic-card-list/components/index.ts
@@ -1,0 +1,1 @@
+export { TopicCardComponent } from './topic-card/topic-card.component';
