+++ packages/frontend/src/features/quiz/ui/topic-card-list/components/topic-card/topic-card.component.scss
