+++ packages/frontend/src/features/quiz/ui/topic-card-list/topic-card-list.types.ts
@@ -1,6 +1,6 @@
 interface TopicCardListItem {
   id: string;
-  topic: string;
+  name: string;
   description: string;
   questionsCount: number;
   score?: number;
